Refactor GoogleSheets service.
