+++ functions/src/modules/SQR.functions.ts
@@ -308,15 +308,15 @@ const parseAudioChunkRemark = (string) => {
   let matches = regex.exec(string);
   while (matches) {
     tokens.push({
-      entireFile: matches[2]? true : null,
-      beginning: matches[2]? null : matches[3],
-      ending: matches[2]? null : matches[4],
+      entireFile: matches[2] ? true : null,
+      beginning: matches[2] ? null : matches[3],
+      ending: matches[2] ? null : matches[4],
       type: matches[5].trim(),
       description: matches[6].trim()
     });
     matches = regex.exec(string);
   }
-  
+
   return tokens;
 }
 /////////////////////////////////////////////////
@@ -334,7 +334,8 @@ export const importSpreadSheetData = functions.https.onRequest(
     ////////////////////////
     //     Submissions
     ////////////////////////
-    const submissionsSheet = new GoogleSheets(spreadsheetId, ISoundQualityReportSheet.Submissions);
+    const submissionsSpreadsheet = new GoogleSheets(spreadsheetId);
+    const submissionsSheet = await submissionsSpreadsheet.useSheet(ISoundQualityReportSheet.Submissions);
     const submissionsRows = await submissionsSheet.getRows();
     for (const row of submissionsRows) {
       const audioFileName = row['Audio File Name'];
@@ -343,7 +344,7 @@ export const importSpreadSheetData = functions.https.onRequest(
       const token = /.*token=([\w-]+)/.exec(row['Update Link'])[1];
 
       const soundIssues = parseAudioChunkRemark(row['Sound Issues']);
-      const unwantedParts = parseAudioChunkRemark(row['Unwanted Parts']); 
+      const unwantedParts = parseAudioChunkRemark(row['Unwanted Parts']);
 
       const submission = {
         completed: row['Completed'] || null,
@@ -364,7 +365,8 @@ export const importSpreadSheetData = functions.https.onRequest(
     //     Allotments
     ////////////////////////
 
-    const allotmentsSheet = new GoogleSheets(spreadsheetId, ISoundQualityReportSheet.Allotments);
+    const allotmentsSpreadsheet = new GoogleSheets(spreadsheetId);
+    const allotmentsSheet = await allotmentsSpreadsheet.useSheet(ISoundQualityReportSheet.Allotments);
     const allotmentsRows = await allotmentsSheet.getRows();
 
     for (const row of allotmentsRows) {
@@ -407,11 +409,9 @@ export const exportAllotmentsToSpreadsheet = functions.database
       const { fileName } = context.params;
       const changedValues = change.after.val();
 
-      const gsheets = new GoogleSheets(
-        functions.config().sqr.spreadsheet_id,
-        ISoundQualityReportSheet.Allotments
-      );
-      const allotmentFileNames = await gsheets.getColumn('File Name');
+      const gsheets = new GoogleSheets(functions.config().sqr.spreadsheet_id);
+      const sheet = await gsheets.useSheet(ISoundQualityReportSheet.Allotments);
+      const allotmentFileNames = await sheet.getColumn('File Name');
       const rowNumber = allotmentFileNames.indexOf(fileName) + 1;
       const { languages, notes, soundQualityReporting } = changedValues;
 
@@ -422,7 +422,7 @@ export const exportAllotmentsToSpreadsheet = functions.database
         followUp,
       } = soundQualityReporting;
 
-      const row: any = await gsheets.getRow(rowNumber);
+      const row: any = await sheet.getRow(rowNumber);
       row['Date Given'] = spreadsheetDateFormat(timestampGiven);
       row['Notes'] = withDefault(notes);
       row['Language'] = commaSeparated(languages);
@@ -431,7 +431,7 @@ export const exportAllotmentsToSpreadsheet = functions.database
       row['Email'] = withDefault(assignee.emailAddress);
       row['Date Done'] = spreadsheetDateFormat(timestampDone);
       row['Follow Up'] = withDefault(followUp);
-      await gsheets.updateRow(rowNumber, row);
+      await sheet.updateRow(rowNumber, row);
     }
   );
 
@@ -479,10 +479,9 @@ export const exportSubmissionsToSpreadsheet = functions.database
       snapshot: functions.database.DataSnapshot,
       context: functions.EventContext
     ): Promise<any> => {
-      const gsheets = new GoogleSheets(
-        functions.config().sqr.spreadsheet_id,
-        ISoundQualityReportSheet.Submissions
-      );
+
+      const gsheets = new GoogleSheets(functions.config().sqr.spreadsheet_id)
+      const submissionSheet = await gsheets.useSheet(ISoundQualityReportSheet.Submissions);
       const {
         author,
         changed,
@@ -496,7 +495,7 @@ export const exportSubmissionsToSpreadsheet = functions.database
         unwantedParts,
       } = snapshot.val();
 
-      await gsheets.appendRow({
+      await submissionSheet.appendRow({
         Completed: spreadsheetDateFormat(completed),
         Updated: spreadsheetDateFormat(changed),
         'Submission Serial': context.params.submission_id,
