+++ functions/src/modules/User.functions.ts
@@ -44,12 +44,10 @@ enum Decision {
  */
 export const importUserRegistrationData = functions.https.onRequest(
   async (req: functions.Request, res: functions.Response) => {
-    const gsheets: GoogleSheet = new GoogleSheet(
-      functions.config().registrations.spreadsheet_id,
-      'Registrations'
-    );
+    const gsheets: GoogleSheet = new GoogleSheet(functions.config().registrations.spreadsheet_id);
+    const sheet = await gsheets.useSheet('Registrations');
 
-    const registrationRows = await gsheets.getRows();
+    const registrationRows = await sheet.getRows();
 
 
     const readyForDatabaseUpdate = registrationRows.map((row: any) => {
